se actualizo int a int16

+++ Pelicula/Program.cs
@@ -8,7 +8,7 @@ namespace Pelicula
     {
         //Campos
         public string titulo;
-        public int año;
+        public Int16 año;
         public string pais;
         public string director;
         private List<Actor> actores = new List<Actor>();
